report.sh: also accommodate 'join lookup_...'
In the code on master, lookup_value was hit via a left outer join rather
than a 'select from lookup_value', so my counts were incorrect.

After this adjustment, the report indicates performance of master (in
terms of number of queries for message lookups) is back to where it was
in Mifos 2.0.